#160 code cleanup for api restructure and renaming the router.
